Move tests to within their source files.
It's my fault that we didn't do this in the first place. It meant that we had to
expose nearly every method as public, even when we really didn't want to. Mea
culpa.
