+++ src/lib/stategraph.rs
@@ -22,7 +22,7 @@ use grammar::{AIdx, Grammar, NIdx, Symbol, SIdx, TIdx};
 
 
 /// The type of "context" (also known as "lookaheads")
-pub type Ctx = BitVec;
+type Ctx = BitVec;
 
 /// Firsts stores all the first sets for a given grammar.
 ///
@@ -49,7 +49,7 @@ pub type Ctx = BitVec;
 /// assert!(firsts.is_epsilon_set(grm.nonterminal_off("A")));
 /// ```
 #[derive(Debug)]
-pub struct Firsts {
+struct Firsts {
     // The representation is a contiguous bitfield, of (terms_len * 1) * nonterms_len. Put another
     // way, each nonterminal has (terms_len + 1) bits, where the bit at position terms_len
     // represents epsilon. So for the grammar given above, the bitvector would be two sequences of
@@ -63,7 +63,7 @@ pub struct Firsts {
 
 impl Firsts {
     /// Generates and returns the firsts set for the given grammar.
-    pub fn new(grm: &Grammar) -> Firsts {
+    fn new(grm: &Grammar) -> Firsts {
         let mut alt_firsts = Vec::with_capacity(grm.nonterms_len);
         for _ in 0..grm.nonterms_len {
             alt_firsts.push(BitVec::from_elem(grm.terms_len, false));
@@ -141,17 +141,17 @@ impl Firsts {
     }
 
     /// Returns true if the terminal `tidx` is in the first set for nonterminal `nidx` is set.
-    pub fn is_set(&self, nidx: NIdx, tidx: TIdx) -> bool {
+    fn is_set(&self, nidx: NIdx, tidx: TIdx) -> bool {
         self.alt_firsts[nidx][tidx]
     }
 
     /// Get all the firsts for alternative `nidx` as a `Ctx`.
-    pub fn alt_firsts(&self, nidx: NIdx) -> &Ctx {
+    fn alt_firsts(&self, nidx: NIdx) -> &Ctx {
         &self.alt_firsts[nidx]
     }
 
     /// Returns true if the nonterminal `nidx` has epsilon in its first set.
-    pub fn is_epsilon_set(&self, nidx: NIdx) -> bool {
+    fn is_epsilon_set(&self, nidx: NIdx) -> bool {
         self.alt_epsilons[nidx]
     }
 
@@ -170,19 +170,19 @@ impl Firsts {
 }
 
 #[derive(Clone, Debug, PartialEq)]
-pub struct Itemset {
-    pub items: HashMap<(AIdx, SIdx), Ctx>
+struct Itemset {
+    items: HashMap<(AIdx, SIdx), Ctx>
 }
 
 impl Itemset {
     /// Create a blank Itemset.
-    pub fn new(_: &Grammar) -> Itemset {
+    fn new(_: &Grammar) -> Itemset {
         Itemset {items: HashMap::new()}
     }
 
     /// Add an item `(alt, dot)` with context `ctx` to this itemset. Returns true if this led to
     /// any changes in the itemset.
-    pub fn add(&mut self, alt: AIdx, dot: SIdx, ctx: &Ctx) -> bool {
+    fn add(&mut self, alt: AIdx, dot: SIdx, ctx: &Ctx) -> bool {
         let entry = self.items.entry((alt, dot));
         match entry {
             Entry::Occupied(mut e) => {
@@ -195,7 +195,7 @@ impl Itemset {
         }
     }
 
-    pub fn close(&self, grm: &Grammar, firsts: &Firsts) -> Itemset {
+    fn close(&self, grm: &Grammar, firsts: &Firsts) -> Itemset {
         // This function can be seen as a merger of getClosure and getContext from Chen's
         // dissertation. It is written in such a style that it allocates relatively little memory
         // and does as few hashmap lookups as practical.
@@ -257,7 +257,7 @@ impl Itemset {
     }
 
     /// Create a new Itemset based on calculating the goto of 'sym' on the current Itemset.
-    pub fn goto(&self, grm: &Grammar, sym: &Symbol) -> Itemset {
+    fn goto(&self, grm: &Grammar, sym: &Symbol) -> Itemset {
         // This is called 'transition' in Chen's dissertation, though note that the definition
         // therein appears to get the dot in the input/output the wrong way around.
         let mut newis = Itemset::new(&grm);
@@ -272,7 +272,7 @@ impl Itemset {
     }
 
     /// Return true if `other` is weakly compatible with `self`.
-    pub fn weakly_compatible(&self, other: &Itemset) -> bool {
+    fn weakly_compatible(&self, other: &Itemset) -> bool {
         // The weakly compatible check is one of the three core parts of Pager's algorithm
         // (along with merging and change propagation). In essence, there are three conditions
         // which, if satisfied, guarantee that `other` is weakly compatible with `self`
@@ -336,7 +336,7 @@ impl Itemset {
 
     /// Merge `other` into `self`, returning `true` if this led to any changes. If `other` is not
     /// weakly compatible with `self`, this function's effects and return value are undefined.
-    pub fn weakly_merge(&mut self, other: &Itemset) -> bool {
+    fn weakly_merge(&mut self, other: &Itemset) -> bool {
         let mut changed = false;
         for (&(alt_i, dot), ctx) in self.items.iter_mut() {
             if ctx.union(&other.items[&(alt_i, dot)]) {
@@ -358,7 +358,7 @@ fn bitvec_intersect(v1: &BitVec, v2: &BitVec) -> bool {
 }
 
 pub struct StateGraph {
-    pub states: Vec<Itemset>,
+    states: Vec<Itemset>,
     pub edges: HashMap<(usize, Symbol), usize>
 }
 
@@ -481,7 +481,9 @@ mod test {
     extern crate bit_vec;
     use self::bit_vec::BitVec;
 
-    use super::bitvec_intersect;
+    use super::{bitvec_intersect, Itemset, Firsts, StateGraph};
+    use grammar::{AIdx, ast_to_grammar, Grammar, Symbol};
+    use yacc::parse_yacc;
 
     #[test]
     fn test_bitvec_intersect() {
@@ -507,5 +509,497 @@ mod test {
         }
         assert!(bitvec_intersect(&b1, &b2));
     }
-}
 
+    fn has(grm: &Grammar, firsts: &Firsts, rn: &str, should_be: Vec<&str>) {
+        let nt_i = grm.nonterminal_off(rn);
+        for (i, n) in grm.terminal_names.iter().enumerate() {
+            match should_be.iter().position(|x| x == n) {
+                Some(_) => {
+                    if !firsts.is_set(nt_i, i) {
+                        panic!("{} is not set in {}", n, rn);
+                    }
+                }
+                None    => {
+                    if firsts.is_set(nt_i, i) {
+                        panic!("{} is incorrectly set in {}", n, rn);
+                    }
+                }
+            }
+        }
+        if should_be.iter().position(|x| x == &"".to_string()).is_some() {
+            assert!(firsts.is_epsilon_set(nt_i));
+        }
+    }
+
+    #[test]
+    fn test_first(){
+        let ast = parse_yacc(&"
+          %start C
+          %token c d
+          %%
+          C: 'c';
+          D: 'd';
+          E: D | C;
+          F: E;
+          ".to_string()).unwrap();
+        let grm = ast_to_grammar(&ast);
+        let firsts = Firsts::new(&grm);
+        has(&grm, &firsts, "^", vec!["c"]);
+        has(&grm, &firsts, "D", vec!["d"]);
+        has(&grm, &firsts, "E", vec!["d", "c"]);
+        has(&grm, &firsts, "F", vec!["d", "c"]);
+    }
+
+    #[test]
+    fn test_first_no_subsequent_nonterminals() {
+        let ast = parse_yacc(&"
+          %start C
+          %token c d
+          %%
+          C: 'c';
+          D: 'd';
+          E: D C;
+          ".to_string()).unwrap();
+        let grm = ast_to_grammar(&ast);
+        let firsts = Firsts::new(&grm);
+        has(&grm, &firsts, "E", vec!["d"]);
+    }
+
+    #[test]
+    fn test_first_epsilon() {
+        let ast = parse_yacc(&"
+          %start A
+          %token a b c
+          %%
+          A: B 'a';
+          B: 'b' | ;
+          C: 'c' | ;
+          D: C;
+          ".to_string()).unwrap();
+        let grm = ast_to_grammar(&ast);
+        let firsts = Firsts::new(&grm);
+        has(&grm, &firsts, "A", vec!["b", "a"]);
+        has(&grm, &firsts, "C", vec!["c", ""]);
+        has(&grm, &firsts, "D", vec!["c", ""]);
+    }
+
+    #[test]
+    fn test_last_epsilon() {
+        let ast = parse_yacc(&"
+          %start A
+          %token b c
+          %%
+          A: B C;
+          B: 'b' | ;
+          C: B 'c' B;
+          ".to_string()).unwrap();
+        let grm = ast_to_grammar(&ast);
+        let firsts = Firsts::new(&grm);
+        has(&grm, &firsts, "A", vec!["b", "c"]);
+        has(&grm, &firsts, "B", vec!["b", ""]);
+        has(&grm, &firsts, "C", vec!["b", "c"]);
+    }
+
+    #[test]
+    fn test_first_no_multiples() {
+        let ast = parse_yacc(&"
+          %start A
+          %token b c
+          %%
+          A: B 'b';
+          B: 'b' | ;
+          ".to_string()).unwrap();
+        let grm = ast_to_grammar(&ast);
+        let firsts = Firsts::new(&grm);
+        has(&grm, &firsts, "A", vec!["b"]);
+    }
+
+    fn eco_grammar() -> Grammar {
+        let ast = parse_yacc(&"
+          %start S
+          %token a b c d f
+          %%
+          S: S 'b' | 'b' A 'a' | 'a';
+          A: 'a' S 'c' | 'a' | 'a' S 'b';
+          B: A S;
+          C: D A;
+          D: 'd' | ;
+          F: C D 'f';
+          ".to_string()).unwrap();
+        ast_to_grammar(&ast)
+    }
+
+    #[test]
+    fn test_first_from_eco() {
+        let grm = eco_grammar();
+        let firsts = Firsts::new(&grm);
+        has(&grm, &firsts, "S", vec!["a", "b"]);
+        has(&grm, &firsts, "A", vec!["a"]);
+        has(&grm, &firsts, "B", vec!["a"]);
+        has(&grm, &firsts, "D", vec!["d", ""]);
+        has(&grm, &firsts, "C", vec!["d", "a"]);
+        has(&grm, &firsts, "F", vec!["d", "a"]);
+    }
+
+    #[test]
+    fn test_first_from_eco_bug() {
+        let ast = parse_yacc(&"
+          %start E
+          %token a b c d e f
+          %%
+          E : T | E 'b' T;
+          T : P | T 'e' P;
+          P : 'a';
+          C: C 'c' | ;
+          D: D 'd' | F;
+          F: 'f' | ;
+          G: C D;
+          ".to_string()).unwrap();
+        let grm = ast_to_grammar(&ast);
+        let firsts = Firsts::new(&grm);
+        has(&grm, &firsts, "E", vec!["a"]);
+        has(&grm, &firsts, "T", vec!["a"]);
+        has(&grm, &firsts, "P", vec!["a"]);
+        has(&grm, &firsts, "C", vec!["c", ""]);
+        has(&grm, &firsts, "D", vec!["f", "d", ""]);
+        has(&grm, &firsts, "G", vec!["c", "d", "f", ""]);
+    }
+
+    fn state_exists(grm: &Grammar, is: &Itemset, nt: &str, alt_off: AIdx, dot: usize, la:
+                        Vec<&str>) {
+        let ab_alt_off = grm.rules_alts[grm.nonterminal_off(nt)][alt_off];
+        let ctx = &is.items[&(ab_alt_off, dot)];
+        for i in 0..grm.terms_len {
+            let bit = ctx[i];
+            let mut found = false;
+            for t in la.iter() {
+                if grm.terminal_off(t) == i {
+                    if !bit {
+                        panic!("bit for terminal {}, dot {} is not set in alternative {} of {} when it should be",
+                               t, dot, alt_off, nt);
+                    }
+                    found = true;
+                    break;
+                }
+            }
+            if !found && bit {
+                panic!("bit for terminal {}, dot {} is set in alternative {} of {} when it shouldn't be",
+                       grm.terminal_names[i], dot, alt_off, nt);
+            }
+        }
+    }
+
+    fn num_active_states(is: &Itemset) -> usize {
+        is.items.len()
+    }
+
+
+    #[test]
+    fn test_dragon_grammar() {
+        // From http://binarysculpting.com/2012/02/04/computing-lr1-closure/
+        let grm = ast_to_grammar(&parse_yacc(&"
+          %start S
+          %%
+          S: L '=' R | R;
+          L: '*' R | 'id';
+          R: L;
+          ".to_string()).unwrap());
+        let firsts = Firsts::new(&grm);
+
+        let mut is = Itemset::new(&grm);
+        let mut la = BitVec::from_elem(grm.terms_len, false);
+        la.set(grm.terminal_off("$"), true);
+        is.add(grm.rules_alts[grm.nonterminal_off("^") as usize][0], 0, &la);
+        let cls_is = is.close(&grm, &firsts);
+        println!("{:?}", cls_is);
+        assert_eq!(num_active_states(&cls_is), 6);
+        state_exists(&grm, &cls_is, "^", 0, 0, vec!["$"]);
+        state_exists(&grm, &cls_is, "S", 0, 0, vec!["$"]);
+        state_exists(&grm, &cls_is, "S", 1, 0, vec!["$"]);
+        state_exists(&grm, &cls_is, "L", 0, 0, vec!["$", "="]);
+        state_exists(&grm, &cls_is, "L", 1, 0, vec!["$", "="]);
+        state_exists(&grm, &cls_is, "R", 0, 0, vec!["$"]);
+    }
+
+    #[test]
+    fn test_closure1_ecogrm() {
+        let grm = eco_grammar();
+        let firsts = Firsts::new(&grm);
+
+        let mut is = Itemset::new(&grm);
+        let mut la = BitVec::from_elem(grm.terms_len, false);
+        la.set(grm.terminal_off("$"), true);
+        is.add(grm.rules_alts[grm.nonterminal_off("^") as usize][0], 0, &la);
+        let mut cls_is = is.close(&grm, &firsts);
+
+        state_exists(&grm, &cls_is, "^", 0, 0, vec!["$"]);
+        state_exists(&grm, &cls_is, "S", 0, 0, vec!["b", "$"]);
+        state_exists(&grm, &cls_is, "S", 1, 0, vec!["b", "$"]);
+        state_exists(&grm, &cls_is, "S", 2, 0, vec!["b", "$"]);
+
+        is = Itemset::new(&grm);
+        is.add(grm.rules_alts[grm.nonterminal_off("F") as usize][0], 0, &la);
+        cls_is = is.close(&grm, &firsts);
+        state_exists(&grm, &cls_is, "F", 0, 0, vec!["$"]);
+        state_exists(&grm, &cls_is, "C", 0, 0, vec!["d", "f"]);
+        state_exists(&grm, &cls_is, "D", 0, 0, vec!["a"]);
+        state_exists(&grm, &cls_is, "D", 1, 0, vec!["a"]);
+    }
+
+    // GrammarAST from 'LR(k) Analyse fuer Pragmatiker'
+    // Z : S
+    // S : Sb
+    //     bAa
+    // A : aSc
+    //     a
+    //     aSb
+    fn grammar3() -> Grammar {
+        ast_to_grammar(&parse_yacc(&"
+          %start S
+          %token a b c d
+          %%
+          S: S 'b' | 'b' A 'a';
+          A: 'a' S 'c' | 'a' | 'a' S 'b';
+          ".to_string()).unwrap())
+    }
+
+    #[test]
+    fn test_closure1_grm3() {
+        let grm = grammar3();
+        let firsts = Firsts::new(&grm);
+
+        let mut is = Itemset::new(&grm);
+        let mut la = BitVec::from_elem(grm.terms_len, false);
+        la.set(grm.terminal_off("$"), true);
+        is.add(grm.rules_alts[grm.nonterminal_off("^") as usize][0], 0, &la);
+        let mut cls_is = is.close(&grm, &firsts);
+
+        state_exists(&grm, &cls_is, "^", 0, 0, vec!["$"]);
+        state_exists(&grm, &cls_is, "S", 0, 0, vec!["b", "$"]);
+        state_exists(&grm, &cls_is, "S", 1, 0, vec!["b", "$"]);
+
+        is = Itemset::new(&grm);
+        la = BitVec::from_elem(grm.terms_len, false);
+        la.set(grm.terminal_off("b"), true);
+        la.set(grm.terminal_off("$"), true);
+        is.add(grm.rules_alts[grm.nonterminal_off("S") as usize][1], 1, &la);
+        cls_is = is.close(&grm, &firsts);
+        state_exists(&grm, &cls_is, "A", 0, 0, vec!["a"]);
+        state_exists(&grm, &cls_is, "A", 1, 0, vec!["a"]);
+        state_exists(&grm, &cls_is, "A", 2, 0, vec!["a"]);
+
+        is = Itemset::new(&grm);
+        la = BitVec::from_elem(grm.terms_len, false);
+        la.set(grm.terminal_off("a"), true);
+        is.add(grm.rules_alts[grm.nonterminal_off("A") as usize][0], 1, &la);
+        cls_is = is.close(&grm, &firsts);
+        state_exists(&grm, &cls_is, "S", 0, 0, vec!["b", "c"]);
+        state_exists(&grm, &cls_is, "S", 1, 0, vec!["b", "c"]);
+    }
+
+    #[test]
+    fn test_goto1() {
+        let grm = grammar3();
+        let firsts = Firsts::new(&grm);
+
+        let mut is = Itemset::new(&grm);
+        let mut la = BitVec::from_elem(grm.terms_len, false);
+        la.set(grm.terminal_off("$"), true);
+        is.add(grm.rules_alts[grm.nonterminal_off("^") as usize][0], 0, &la);
+        let cls_is = is.close(&grm, &firsts);
+
+        let goto1 = cls_is.goto(&grm, &Symbol::Nonterminal(grm.nonterminal_off("S")));
+        state_exists(&grm, &goto1, "^", 0, 1, vec!["$"]);
+        state_exists(&grm, &goto1, "S", 0, 1, vec!["$", "b"]);
+
+        // follow 'b' from start set
+        let goto2 = cls_is.goto(&grm, &Symbol::Terminal(grm.terminal_off("b")));
+        state_exists(&grm, &goto2, "S", 1, 1, vec!["$", "b"]);
+
+        // continue by following 'a' from last goto, after it's been closed
+        let goto3 = goto2.close(&grm, &firsts).goto(&grm, &Symbol::Terminal(grm.terminal_off("a")));
+        state_exists(&grm, &goto3, "A", 1, 1, vec!["a"]);
+        state_exists(&grm, &goto3, "A", 2, 1, vec!["a"]);
+    }
+
+    #[test]
+    fn test_stategraph() {
+        let grm = grammar3();
+        let sg = StateGraph::new(&grm);
+        for st in sg.states.iter() { println!("  {:?}", st); }
+
+        assert_eq!(sg.states.len(), 10);
+        assert_eq!(sg.edges.len(), 10);
+
+        assert_eq!(num_active_states(&sg.states[0]), 1);
+        state_exists(&grm, &sg.states[0], "^", 0, 0, vec!["$"]);
+
+        let s1 = sg.edges[&(0, Symbol::Nonterminal(grm.nonterminal_off("S")))];
+        assert_eq!(num_active_states(&sg.states[s1]), 2);
+        state_exists(&grm, &sg.states[s1], "^", 0, 1, vec!["$"]);
+        state_exists(&grm, &sg.states[s1], "S", 0, 1, vec!["$", "b"]);
+
+        let s2 = sg.edges[&(s1, Symbol::Terminal(grm.terminal_off("b")))];
+        assert_eq!(num_active_states(&sg.states[s2]), 1);
+        state_exists(&grm, &sg.states[s2], "S", 0, 2, vec!["$", "b"]);
+
+        let s3 = sg.edges[&(0, Symbol::Terminal(grm.terminal_off("b")))];
+        assert_eq!(num_active_states(&sg.states[s3]), 1);
+        state_exists(&grm, &sg.states[s3], "S", 1, 1, vec!["$", "b", "c"]);
+
+        let s4 = sg.edges[&(s3, Symbol::Nonterminal(grm.nonterminal_off("A")))];
+        assert_eq!(num_active_states(&sg.states[s4]), 1);
+        state_exists(&grm, &sg.states[s4], "S", 1, 2, vec!["$", "b", "c"]);
+
+        let s5 = sg.edges[&(s4, Symbol::Terminal(grm.terminal_off("a")))];
+        assert_eq!(num_active_states(&sg.states[s5]), 1);
+        state_exists(&grm, &sg.states[s5], "S", 1, 3, vec!["$", "b", "c"]);
+
+        let s6 = sg.edges[&(s3, Symbol::Terminal(grm.terminal_off("a")))];
+        assert_eq!(s3, sg.edges[&(s6, Symbol::Terminal(grm.terminal_off("b")))]); // result from merging 10 into 3
+        assert_eq!(num_active_states(&sg.states[s6]), 3);
+        state_exists(&grm, &sg.states[s6], "A", 0, 1, vec!["a"]);
+        state_exists(&grm, &sg.states[s6], "A", 1, 1, vec!["a"]);
+        state_exists(&grm, &sg.states[s6], "A", 2, 1, vec!["a"]);
+
+        let s7 = sg.edges[&(s6, Symbol::Nonterminal(grm.nonterminal_off("S")))];
+        assert_eq!(num_active_states(&sg.states[s7]), 3);
+        state_exists(&grm, &sg.states[s7], "A", 0, 2, vec!["a"]);
+        state_exists(&grm, &sg.states[s7], "A", 2, 2, vec!["a"]);
+        state_exists(&grm, &sg.states[s7], "S", 0, 1, vec!["b", "c"]);
+
+        let s8 = sg.edges[&(s7, Symbol::Terminal(grm.terminal_off("c")))];
+        assert_eq!(num_active_states(&sg.states[s8]), 1);
+        state_exists(&grm, &sg.states[s8], "A", 0, 3, vec!["a"]);
+
+        let s9 = sg.edges[&(s7, Symbol::Terminal(grm.terminal_off("b")))];
+        assert_eq!(num_active_states(&sg.states[s9]), 2);
+        state_exists(&grm, &sg.states[s9], "A", 2, 3, vec!["a"]);
+        state_exists(&grm, &sg.states[s9], "S", 0, 2, vec!["b", "c"]);
+    }
+
+    // Pager grammar
+    fn grammar_pager() -> Grammar {
+        ast_to_grammar(&parse_yacc(&"
+            %start X
+            %%
+             X : 'a' Y 'd' | 'a' Z 'c' | 'a' T | 'b' Y 'e' | 'b' Z 'd' | 'b' T;
+             Y : 't' W | 'u' X;
+             Z : 't' 'u';
+             T : 'u' X 'a';
+             W : 'u' V;
+             V : ;
+          ".to_string()).unwrap())
+    }
+
+    #[test]
+    fn test_pager_graph() {
+        let grm = grammar_pager();
+        let sg = StateGraph::new(&grm);
+
+        assert_eq!(sg.states.len(), 23);
+        assert_eq!(sg.edges.len(), 27);
+
+        // State 0
+        assert_eq!(num_active_states(&sg.states[0]), 1);
+        state_exists(&grm, &sg.states[0], "^", 0, 0, vec!["$"]);
+
+        let s1 = sg.edges[&(0, Symbol::Terminal(grm.terminal_off("a")))];
+        assert_eq!(num_active_states(&sg.states[s1]), 3);
+        state_exists(&grm, &sg.states[s1], "X", 0, 1, vec!["a", "d", "e", "$"]);
+        state_exists(&grm, &sg.states[s1], "X", 1, 1, vec!["a", "d", "e", "$"]);
+        state_exists(&grm, &sg.states[s1], "X", 2, 1, vec!["a", "d", "e", "$"]);
+
+        let s7 = sg.edges[&(0, Symbol::Terminal(grm.terminal_off("b")))];
+        assert_eq!(num_active_states(&sg.states[s7]), 3);
+        state_exists(&grm, &sg.states[s7], "X", 3, 1, vec!["a", "d", "e", "$"]);
+        state_exists(&grm, &sg.states[s7], "X", 4, 1, vec!["a", "d", "e", "$"]);
+        state_exists(&grm, &sg.states[s7], "X", 5, 1, vec!["a", "d", "e", "$"]);
+
+        let s4 = sg.edges[&(s1, Symbol::Terminal(grm.terminal_off("u")))];
+        assert_eq!(num_active_states(&sg.states[s4]), 2);
+        assert_eq!(s4, sg.edges[&(s7, Symbol::Terminal(grm.terminal_off("u")))]);
+        state_exists(&grm, &sg.states[s4], "Y", 1, 1, vec!["d", "e"]);
+        state_exists(&grm, &sg.states[s4], "T", 0, 1, vec!["a", "d", "e", "$"]);
+
+        assert_eq!(s1, sg.edges[&(s4, Symbol::Terminal(grm.terminal_off("a")))]);
+        assert_eq!(s7, sg.edges[&(s4, Symbol::Terminal(grm.terminal_off("b")))]);
+
+        let s2 = sg.edges[&(s1, Symbol::Terminal(grm.terminal_off("t")))];
+        assert_eq!(num_active_states(&sg.states[s2]), 2);
+        state_exists(&grm, &sg.states[s2], "Y", 0, 1, vec!["d"]);
+        state_exists(&grm, &sg.states[s2], "Z", 0, 1, vec!["c"]);
+
+        let s3 = sg.edges[&(s2, Symbol::Terminal(grm.terminal_off("u")))];
+        assert_eq!(num_active_states(&sg.states[s3]), 2);
+        state_exists(&grm, &sg.states[s3], "Z", 0, 2, vec!["c"]);
+        state_exists(&grm, &sg.states[s3], "W", 0, 1, vec!["d"]);
+
+        let s5 = sg.edges[&(s4, Symbol::Nonterminal(grm.nonterminal_off("X")))];
+        assert_eq!(num_active_states(&sg.states[s5]), 2);
+        state_exists(&grm, &sg.states[s5], "Y", 1, 2, vec!["d", "e"]);
+        state_exists(&grm, &sg.states[s5], "T", 0, 2, vec!["a", "d", "e", "$"]);
+
+        let s6 = sg.edges[&(s5, Symbol::Terminal(grm.terminal_off("a")))];
+        assert_eq!(num_active_states(&sg.states[s6]), 1);
+        state_exists(&grm, &sg.states[s6], "T", 0, 3, vec!["a", "d", "e", "$"]);
+        
+        let s8 = sg.edges[&(s7, Symbol::Terminal(grm.terminal_off("t")))];
+        assert_eq!(num_active_states(&sg.states[s8]), 2);
+        state_exists(&grm, &sg.states[s8], "Y", 0, 1, vec!["e"]);
+        state_exists(&grm, &sg.states[s8], "Z", 0, 1, vec!["d"]);
+
+        let s9 = sg.edges[&(s8, Symbol::Terminal(grm.terminal_off("u")))];
+        assert_eq!(num_active_states(&sg.states[s9]), 2);
+        state_exists(&grm, &sg.states[s9], "Z", 0, 2, vec!["d"]);
+        state_exists(&grm, &sg.states[s9], "W", 0, 1, vec!["e"]);
+
+        // Ommitted successors from the graph in Fig.3
+        
+        // X-successor of S0
+        let s0x = sg.edges[&(0, Symbol::Nonterminal(grm.nonterminal_off("X")))];
+        state_exists(&grm, &sg.states[s0x], "^", 0, 1, vec!["$"]);
+
+        // Y-successor of S1 (and it's d-successor)
+        let s1y = sg.edges[&(s1, Symbol::Nonterminal(grm.nonterminal_off("Y")))];
+        state_exists(&grm, &sg.states[s1y], "X", 0, 2, vec!["a", "d", "e", "$"]);
+        let s1yd = sg.edges[&(s1y, Symbol::Terminal(grm.terminal_off("d")))];
+        state_exists(&grm, &sg.states[s1yd], "X", 0, 3, vec!["a", "d", "e", "$"]);
+
+        // Z-successor of S1 (and it's successor)
+        let s1z = sg.edges[&(s1, Symbol::Nonterminal(grm.nonterminal_off("Z")))];
+        state_exists(&grm, &sg.states[s1z], "X", 1, 2, vec!["a", "d", "e", "$"]);
+        let s1zc = sg.edges[&(s1z, Symbol::Terminal(grm.terminal_off("c")))];
+        state_exists(&grm, &sg.states[s1zc], "X", 1, 3, vec!["a", "d", "e", "$"]);
+
+        // T-successor of S1
+        let s1t = sg.edges[&(s1, Symbol::Nonterminal(grm.nonterminal_off("T")))];
+        state_exists(&grm, &sg.states[s1t], "X", 2, 2, vec!["a", "d", "e", "$"]);
+
+        // Y-successor of S7 (and it's d-successor)
+        let s7y = sg.edges[&(s7, Symbol::Nonterminal(grm.nonterminal_off("Y")))];
+        state_exists(&grm, &sg.states[s7y], "X", 3, 2, vec!["a", "d", "e", "$"]);
+        let s7ye = sg.edges[&(s7y, Symbol::Terminal(grm.terminal_off("e")))];
+        state_exists(&grm, &sg.states[s7ye], "X", 3, 3, vec!["a", "d", "e", "$"]);
+
+        // Z-successor of S7 (and it's successor)
+        let s7z = sg.edges[&(s7, Symbol::Nonterminal(grm.nonterminal_off("Z")))];
+        state_exists(&grm, &sg.states[s7z], "X", 4, 2, vec!["a", "d", "e", "$"]);
+        let s7zd = sg.edges[&(s7z, Symbol::Terminal(grm.terminal_off("d")))];
+        state_exists(&grm, &sg.states[s7zd], "X", 4, 3, vec!["a", "d", "e", "$"]);
+
+        // T-successor of S7
+        let s7t = sg.edges[&(s7, Symbol::Nonterminal(grm.nonterminal_off("T")))];
+        state_exists(&grm, &sg.states[s7t], "X", 5, 2, vec!["a", "d", "e", "$"]);
+
+        // W-successor of S2 and S8 (merged)
+        let s8w = sg.edges[&(s8, Symbol::Nonterminal(grm.nonterminal_off("W")))];
+        assert_eq!(s8w, sg.edges[&(s2, Symbol::Nonterminal(grm.nonterminal_off("W")))]);
+        state_exists(&grm, &sg.states[s8w], "Y", 0, 2, vec!["d", "e"]);
+
+        // V-successor of S3 and S9 (merged)
+        let s9v = sg.edges[&(s9, Symbol::Nonterminal(grm.nonterminal_off("V")))];
+        assert_eq!(s9v, sg.edges[&(s3, Symbol::Nonterminal(grm.nonterminal_off("V")))]);
+        state_exists(&grm, &sg.states[s9v], "W", 0, 2, vec!["d", "e"]);
+    }
+}
